refactor: extract attribute args to its own mod
