+++ crates/bevy_auto_plugin_shared/src/lib.rs
@@ -1,12 +1,13 @@
 use crate::type_list::TypeList;
 use crate::util::{PathExt, path_to_string_with_spaces};
-use darling::{FromDeriveInput, FromField, FromMeta, FromVariant};
+use darling::FromMeta;
 use proc_macro2::{Ident, Span, TokenStream as MacroStream, TokenStream};
 use quote::quote;
 use std::collections::HashSet;
 use std::hash::Hash;
-use syn::{Attribute, Generics, Path, Type, Visibility, parse2};
+use syn::{Path, parse2};
 
+pub mod attribute_args;
 pub mod bevy_app_code_gen;
 pub mod flat_file;
 pub mod global;
@@ -36,110 +37,6 @@ impl AutoPluginAttribute {
             Self::RegisterStateType => "auto_register_state_type",
             Self::AddSystem => "auto_add_system",
         }
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Debug, FromField)]
-pub struct GlobalAutoPluginField {
-    ident: Option<Ident>,
-    ty: Type,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, FromVariant)]
-pub struct GlobalAutoPluginVariant {
-    ident: Ident,
-    fields: darling::ast::Fields<GlobalAutoPluginField>,
-}
-
-#[derive(FromDeriveInput, Debug)]
-#[darling(attributes(auto_plugin), forward_attrs, supports(struct_any, enum_any))]
-pub struct GlobalAutoPluginDeriveParams {
-    pub ident: Ident,
-    pub vis: Visibility,
-    pub generics: Generics,
-    pub data: darling::ast::Data<GlobalAutoPluginVariant, GlobalAutoPluginField>,
-    pub attrs: Vec<Attribute>,
-    #[darling(flatten)]
-    pub auto_plugin: GlobalAutoPluginStructOrEnumAttributeParams,
-}
-
-#[derive(FromMeta, Debug, Default)]
-#[darling(derive_syn_parse, default)]
-pub struct GlobalAutoPluginStructOrEnumAttributeParams {
-    #[darling(multiple)]
-    pub generics: Vec<TypeList>,
-    pub impl_plugin_trait: bool,
-    pub impl_generic_auto_plugin_trait: bool,
-    pub impl_generic_plugin_trait: bool,
-}
-
-#[derive(FromMeta, Debug, Default)]
-#[darling(derive_syn_parse, default)]
-pub struct GlobalAutoPluginFnAttributeParams {
-    #[darling(multiple)]
-    pub generics: Vec<TypeList>,
-}
-
-fn hash_global_struct_or_enum_attribute_params(
-    ident: &Ident,
-    params: &GlobalStructOrEnumAttributeParams,
-) -> String {
-    use std::hash::Hasher;
-    let mut hasher = std::collections::hash_map::DefaultHasher::new();
-    ident.hash(&mut hasher);
-    params.hash(&mut hasher);
-    format!("{:x}", hasher.finish())
-}
-
-fn get_unique_ident_string_for_global_struct_or_enum_attribute(
-    prefix: &'static str,
-    ident: &Ident,
-    params: &GlobalStructOrEnumAttributeParams,
-) -> String {
-    let hash = hash_global_struct_or_enum_attribute_params(ident, params);
-    format!("{prefix}_{hash}")
-}
-
-pub fn get_unique_ident_for_global_struct_or_enum_attribute(
-    prefix: &'static str,
-    ident: &Ident,
-    params: &GlobalStructOrEnumAttributeParams,
-) -> Ident {
-    let ident_string =
-        get_unique_ident_string_for_global_struct_or_enum_attribute(prefix, ident, params);
-    Ident::new(&ident_string, ident.span())
-}
-
-#[derive(FromMeta, Debug, PartialEq, Hash)]
-#[darling(derive_syn_parse)]
-pub struct GlobalStructOrEnumAttributeParams {
-    pub plugin: Path,
-    #[darling(flatten, default)]
-    pub inner: StructOrEnumAttributeParams,
-}
-
-impl GlobalStructOrEnumAttributeParams {
-    pub fn has_generics(&self) -> bool {
-        self.inner.has_generics()
-    }
-}
-
-#[derive(FromMeta, Debug, Default, PartialEq, Hash)]
-#[darling(derive_syn_parse, default)]
-pub struct StructOrEnumAttributeParams {
-    // TODO: #[darling(multiple)]
-    //     pub generics: Vec<TypeList>,
-    pub generics: Option<TypeList>,
-}
-
-impl StructOrEnumAttributeParams {
-    pub fn has_generics(&self) -> bool {
-        self.generics
-            .as_ref()
-            .map(|types| !types.0.is_empty())
-            .unwrap_or(false)
     }
 }
